+++ src/main.ts
@@ -3,7 +3,7 @@ import App from './App.vue'
 import router from './router'
 import store from './store'
 import 'view-design/dist/styles/iview.css'
-import ViewUI  from './iview'
+import ViewUI  from './viewui'
 
 Vue.use(ViewUI)
